hero section fixes
- created new JS files and imported them in main.JS for the back-to-top and dark-light-theme functionalities (before these functionalities were in the JS file named hero-search.js)
- fixed dark/light theme pagination issues
- fixed HTML validator issues
