+++ src/index.js
@@ -7,3 +7,5 @@ import './js/buy-tickets.js';
 import './js/modal.js';
 import './js/modal-team.js';
 import './js/loading.js';
+import './js/dark-light-theme.js';
+import './js/back-to-top';
