+++ src/partials/hero.html
@@ -14,8 +14,7 @@
           <input id="search" type="text" name="search" placeholder="Start searching" class="search hero-input"
             autocomplete="off">
           <button id="btn-search" class="btn-search" type="submit">
-            <svg class="btn-search-svg" width="12" height="12" viewBox="0 0 32 32"> fill="none"
-              xmlns="http://www.w3.org/2000/svg">
+            <svg class="btn-search-svg" width="12" height="12" viewBox="0 0 32 32">
               <path
                 d="M4.024 4.024c2.595-2.595 6.045-4.024 9.715-4.024s7.12 1.429 9.715 4.024c2.595 2.595 4.024 6.045 4.024 9.715 0 3.355-1.195 6.526-3.383 9.029l7.63 7.63c0.366 0.366 0.366 0.961 0 1.327-0.183 0.183-0.423 0.275-0.664 0.275s-0.48-0.092-0.663-0.275l-7.63-7.63c-2.502 2.188-5.673 3.383-9.029 3.383-3.67 0-7.12-1.429-9.715-4.024s-4.024-6.045-4.024-9.715 1.429-7.12 4.024-9.715zM5.351 22.127c4.625 4.625 12.15 4.625 16.775 0s4.625-12.15 0-16.775c-2.313-2.313-5.35-3.469-8.388-3.469s-6.076 1.157-8.388 3.469c-4.625 4.625-4.625 12.15 0 16.775z">
               </path>
@@ -51,7 +50,7 @@
   <!-- Dark/Light Mode -->
   <button class="hero_theme-toggle" id="themeToggle">
     <!-- Light Theme Icon -->
-    <svg class="theme-toggle_light-icon" width="24" height="24" id="lightIcon" viewBox="0 0 32 32">>
+    <svg class="theme-toggle_light-icon" width="24" height="24" id="lightIcon" viewBox="0 0 32 32">
       <path
         d="M15.982 24.061c-4.455 0-8.079-3.624-8.079-8.079s3.624-8.079 8.079-8.079c4.455 0 8.079 3.624 8.079 8.079s-3.624 8.079-8.079 8.079z">
       </path>
